New bot and tab (#2)
* TailwindTraders-Bot

Including the bot into the project

* delete bot

* New tab created using admin account

* New bot created using Admin account
