--- teams/TailwindTraders-M365/service/botActivityHandler.js
@@ -1,165 +1,0 @@
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-
-const {
-    TurnContext,
-    MessageFactory,
-    TeamsActivityHandler,
-    CardFactory,
-    ActionTypes
-} = require('botbuilder');
-
-class BotActivityHandler extends TeamsActivityHandler {
-    constructor() {
-        super();
-        /* Conversation Bot */
-        /*  Teams bots are Microsoft Bot Framework bots.
-            If a bot receives a message activity, the turn handler sees that incoming activity
-            and sends it to the onMessage activity handler.
-            Learn more: https://aka.ms/teams-bot-basics.
-
-            NOTE:   Ensure the bot endpoint that services incoming conversational bot queries is
-                    registered with Bot Framework.
-                    Learn more: https://aka.ms/teams-register-bot. 
-        */
-        // Registers an activity event handler for the message event, emitted for every incoming message activity.
-        this.onMessage(async (context, next) => {
-            TurnContext.removeRecipientMention(context.activity);
-            switch (context.activity.text.trim()) {
-            case 'Hello':
-                await this.mentionActivityAsync(context);
-                break;
-            default:
-                // By default for unknown activity sent by user show
-                // a card with the available actions.
-                const value = { count: 0 };
-                const card = CardFactory.heroCard(
-                    'Lets talk...',
-                    null,
-                    [{
-                        type: ActionTypes.MessageBack,
-                        title: 'Say Hello',
-                        value: value,
-                        text: 'Hello'
-                    }]);
-                await context.sendActivity({ attachments: [card] });
-                break;
-            }
-            await next();
-        });
-        /* Conversation Bot */
-    }
-
-    /* Conversation Bot */
-    /**
-     * Say hello and @ mention the current user.
-     */
-    async mentionActivityAsync(context) {
-        const TextEncoder = require('html-entities').XmlEntities;
-
-        const mention = {
-            mentioned: context.activity.from,
-            text: `<at>${ new TextEncoder().encode(context.activity.from.name) }</at>`,
-            type: 'mention'
-        };
-
-        const replyActivity = MessageFactory.text(`Hi ${ mention.text }`);
-        replyActivity.entities = [mention];
-        
-        await context.sendActivity(replyActivity);
-    }
-    /* Conversation Bot */
-
-    /* Messaging Extension - Action */
-    /* Building a messaging extension action command is a two-step process.
-        (1) Define how the messaging extension will look and be invoked in the client.
-            This can be done from the Configuration tab, or the Manifest Editor.
-            Learn more: https://aka.ms/teams-me-design-action.
-        (2) Define how the bot service will respond to incoming action commands.
-            Learn more: https://aka.ms/teams-me-respond-action.
-        
-        NOTE:   Ensure the bot endpoint that services incoming messaging extension queries is
-                registered with Bot Framework.
-                Learn more: https://aka.ms/teams-register-bot.
-    */
-
-    // Invoked when the service receives an incoming action command.
-    handleTeamsMessagingExtensionSubmitAction(context, action) {
-        /* Commands are defined in the manifest file. This can be done using the Configuration tab, editing
-           the manifest Json directly, or using the manifest editor in App Studio. This project includes two
-           commands to help get you started: createCard and shareMessage.
-        */
-        switch (action.commandId) {
-        case 'createCard':
-            return createCardCommand(context, action);
-        case 'shareMessage':
-            return shareMessageCommand(context, action);
-        default:
-            throw new Error('NotImplemented');
-        }
-    }
-    /* Messaging Extension - Action */
-
-}
-
-/* Messaging Extension - Action */
-function createCardCommand(context, action) {
-    // The user has chosen to create a card by choosing the 'Create Card' context menu command.
-    const data = action.data;
-    const heroCard = CardFactory.heroCard(data.title, data.text);
-    heroCard.content.subtitle = data.subTitle;
-    const attachment = { contentType: heroCard.contentType, content: heroCard.content, preview: heroCard };
-
-    return {
-        composeExtension: {
-            type: 'result',
-            attachmentLayout: 'list',
-            attachments: [
-                attachment
-            ]
-        }
-    };
-}
-
-function shareMessageCommand(context, action) {
-    // The user has chosen to share a message by choosing the 'Share Message' context menu command.
-    let userName = 'unknown';
-    if (action.messagePayload.from &&
-            action.messagePayload.from.user &&
-            action.messagePayload.from.user.displayName) {
-        userName = action.messagePayload.from.user.displayName;
-    }
-
-    // This Messaging Extension example allows the user to check a box to include an image with the
-    // shared message.  This demonstrates sending custom parameters along with the message payload.
-    let images = [];
-    const includeImage = action.data.includeImage;
-    if (includeImage === 'true') {
-        images = ['https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcQtB3AwMUeNoq4gUBGe6Ocj8kyh3bXa9ZbV7u1fVKQoyKFHdkqU'];
-    }
-    const heroCard = CardFactory.heroCard(`${ userName } originally sent this message:`,
-        action.messagePayload.body.content,
-        images);
-
-    if (action.messagePayload.attachments && action.messagePayload.attachments.length > 0) {
-        // This sample does not add the MessagePayload Attachments.  This is left as an
-        // exercise for the user.
-        heroCard.content.subtitle = `(${ action.messagePayload.attachments.length } Attachments not included)`;
-    }
-
-    const attachment = { contentType: heroCard.contentType, content: heroCard.content, preview: heroCard };
-
-    return {
-        composeExtension: {
-            type: 'result',
-            attachmentLayout: 'list',
-            attachments: [
-                attachment
-            ]
-        }
-    };
-}
-/* Messaging Extension - Action */
-
-module.exports.BotActivityHandler = BotActivityHandler;
-
