+++ teams/TailwindTraders-M365/tabs/Learn Together Tab/README.md
@@ -1,0 +1,27 @@
+# Tabs
+
+Tabs are Teams-aware webpages embedded in Microsoft Teams. A channel/group tab delivers content to channels and group chats, and are a great way to create collaborative spaces around dedicated web-based content.
+
+## Prerequisites
+-  [NodeJS](https://nodejs.org/en/)
+
+-  [M365 developer account](https://docs.microsoft.com/en-us/microsoftteams/platform/concepts/build-and-test/prepare-your-o365-tenant) or access to a Teams account with the appropriate permissions to install an app.
+
+## Build and Run
+
+In the project directory, execute:
+
+`npm install`
+
+`npm start`
+
+## Deploy to Teams
+Start debugging the project by hitting the `F5` key or click the debug icon in Visual Studio Code and click the `Start Debugging` green arrow button.
+
+### NOTE: First time debug step
+On the first time running and debugging your app you need allow the localhost certificate.  After starting debugging when Chrome is launched and you have installed your app it will fail to load.
+
+- Open a new tab `in the same browser window that was opened`
+- Navigate to `https://localhost:3000/tab`
+- Click the `Advanced` button
+- Select the `Continue to localhost`
